MCU8MASS-1790 Migrate to having cryptoauthlib in a folder and include scripts for updating include path of cryptoauthlib

+++ scripts/inject_cryptoauthlib.sh
@@ -2,6 +2,7 @@
 
 SCRIPTPATH="$( cd "$(dirname "$0")" ; pwd -P )"
 CRYPTOAUTH_PATH=$SCRIPTPATH/../lib/cryptoauth
+TEMPDIR_PATH=$CRYPTOAUTH_PATH/temp
 SRC_PATH=$SCRIPTPATH/../src
 
 # Update cryptoauthlib
@@ -15,28 +16,37 @@ else
     git clone https://github.com/MicrochipTech/cryptoauthlib.git $CRYPTOAUTH_PATH/cryptoauthlib
 fi
 
-echo -n "Injecting... "
+echo "Injecting... "
+
+# Fix the include paths so that all are relative to src since Arduino doesn't 
+# allow us to add extra include paths for the compiler
+python "$SCRIPTPATH/update_include_paths.py" "$CRYPTOAUTH_PATH/cryptoauthlib" "$TEMPDIR_PATH"
+
+# Prevent removing items if the temporary directory was not created, just to be 
+# safe and not removing files from the project
+if [ ! -d "$TEMPDIR_PATH" ]; then
+    echo "Temporary directory for cryptoauthlib not created, aborting"
+    exit 1
+else 
+    # Remove everything we don't need, so we are only left with .h files
+    pushd "$TEMPDIR_PATH/lib" 1> /dev/null 
+        rm -rf mbedtls pkcs11 openssl wolfssl jwt cmake CMakeLists.txt
+
+        pushd hal 1> /dev/null 
+            find . ! -name "atca_hal.*" -type f -exec rm -f {} +
+        popd 1> /dev/null
+    popd 1> /dev/null
+fi
+
+# Make the destination folders
+mkdir "$SRC_PATH/cryptoauthlib"
+mkdir "$SRC_PATH/cryptoauthlib/app"
 
 # Copy sources
-cp -r "$CRYPTOAUTH_PATH/cryptoauthlib/lib/" "$SRC_PATH/cryptoauthlib/"
+cp -r "$TEMPDIR_PATH/lib" "$SRC_PATH/cryptoauthlib/"
+cp -r "$TEMPDIR_PATH/app/tng"  "$SRC_PATH/cryptoauthlib/app/"
 cp -r "$CRYPTOAUTH_PATH/atca_config.h" "$SRC_PATH/cryptoauthlib/"
 
-# Remove everything we don't need, so we are only left with .h files
-pushd "$SRC_PATH/cryptoauthlib" 1> /dev/null 
-    rm -rf mbedtls pkcs11 openssl wolfssl jwt cmake CMakeLists.txt
-
-    pushd hal 1> /dev/null 
-        find . ! -name "atca_hal.*" -type f -exec rm -f {} +
-    popd 1> /dev/null
-popd 1> /dev/null
-
-cp -r "$CRYPTOAUTH_PATH/cryptoauthlib/app/tng/"*.c  "$SRC_PATH/cryptoauthlib/"
-cp -r "$CRYPTOAUTH_PATH/cryptoauthlib/app/tng/"*.h  "$SRC_PATH/cryptoauthlib/"
-
-# Move everything to src since we can't add other include paths for 
-# Arduino, has to be top level in src
-mv "$SRC_PATH/cryptoauthlib/"* "$SRC_PATH" 1> /dev/null
-
-rm -r "$SRC_PATH/cryptoauthlib"
+rm -r "$TEMPDIR_PATH"
 
 echo "Done!"
